question1/pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.8.4",
    "diffusers>=0.37.1",
    "matplotlib>=3.10.8",
    "onnx>=1.19.1",
    "onnxruntime-gpu>=1.22.0",
    "optuna>=4.8.0",
    "pandas>=3.0.2",
    "peft>=0.19.1",
    "pillow>=11.3.0",
    "python-dotenv>=1.2.2",
    "seaborn>=0.13.2",
    "torch==2.6.0",
    "transformers>=5.5.4",
    "wandb>=0.26.0",
]
